requirements.txt
openai>=1.3.0
python-dotenv>=1.0.0
fastapi>=0.111.0
uvicorn>=0.30.0
duckduckgo-search>=6.2.0
requests>=2.31.0
apscheduler>=3.10.4
python-multipart>=0.0.9
pypdf>=4.2.0
python-docx>=1.1.2
stripe>=10.10.0
